refactor(package): view로 패키지 이동
